react-navigation upgrade progress
