+++ plugins/react-navigation/examples/demo/src/rematch/models.js
@@ -1,0 +1,13 @@
+export const auth = {
+  state: {
+    isLoggedIn: false,
+  },
+  reducers: {
+    login: () => ({
+      isLoggedIn: true,
+    }),
+    logout: () => ({
+      isLoggedIn: false,
+    }),
+  }
+}
