+++ plugins/react-navigation/src/redux.ts
@@ -1,0 +1,30 @@
+import {
+  createReactNavigationReduxMiddleware,
+  createReduxBoundAddListener,
+} from 'react-navigation-redux-helpers'
+
+export default ({ AppNavigator, initialScreen }) => {
+  const { router } = AppNavigator
+  const initialState = router.getStateForAction(router.getActionForPathAndParams(initialScreen))
+
+  // reducer
+  const navReducer = (state = initialState, action) => {
+    const nextState = router.getStateForAction(action, state)
+    // Simply return the original `state` if `nextState` is null or undefined.
+    return nextState || state
+  }
+
+  // middleware
+  const navMiddleware = createReactNavigationReduxMiddleware(
+    'root',
+    (state) => state.nav,
+  )
+
+  const addListener = createReduxBoundAddListener('root')
+
+  return {
+    addListener,
+    navMiddleware,
+    navReducer,
+  }
+}
